Avoid eventstore::AppendToStreamOptions
This should only be used within the Kurrent adapter. We only need to
deal with our own EventStreamVersion as an expected version.
